Support "ZeroItem" in custom supplementary views
Currently in the default view model provider the data source provides a
CollectionZeroItemViewModel when it encounters a
CollectionZeroItemModel. This wasn't the case if the implementor
provides a custom view model binding provider leaving it to implement
the exact same "empty" pieces itself (The CollectionZeroItemViewModel
is internal to Pilot).

This code just uses the same logic as the primary path to shortcut
CollectionZeroItemModels and return a CollectionZeroItemViewModel.